ytApi/requirements.txt
django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.5.0
google-api-python-client==1.10.0
requests==2.24.0
ipython==7.16.1
prettytable==0.7.2
# For the persistence stores
psycopg2-binary==2.7.7
dj-database-url==0.5.0
